rewrite algorithm

+++ problem_1.c
@@ -2,16 +2,16 @@
 #include <math.h>
 
 /**
- * @param {double} (i.e. compute sum of n + (n + step))
- * @param {double} max
- * @return {double} sum of all consecutive integers within range
+ * @param {int} (i.e. compute sum of n + (n + step))
+ * @param {int} max
+ * @return {int} sum of all consecutive integers within range
  */
-double consecutiveSum(double step, double max) {
+int consecutiveSum(int step, int max) {
 	max = floor(max / step);
 	return step * floor((pow(max,2) + max) / 2);
 } 
 
 int main() {
-	printf("%f \n",consecutiveSum(3, 999) + consecutiveSum(5, 999) - consecutiveSum(15, 999));
+	printf("%d \n",consecutiveSum(3, 999) + consecutiveSum(5, 999) - consecutiveSum(15, 999));
 	return 0;
 }
